decor_dataset/requirements.txt
flask
flask-cors
scikit-learn
joblib
numpy
pillow
